refactor: remove lib package (#14)
refactor: move `clippy` lints to workspace `Cargo.toml`
