+++ Cargo.toml
@@ -1,8 +1,16 @@
 [workspace]
-
+resolver = "2"
 members = [
     "ashen",
     "engine",
-    "lib"
 ]
-resolver = "2"
+
+[workspace.lints.rust]
+unused = "allow"
+unused_imports = "warn"
+
+[workspace.lints.clippy]
+pedantic = "warn"
+default_trait_access = "allow"
+module_name_repetitions = "allow"
+wildcard_imports = "allow"
