+++ ashen/Cargo.toml
@@ -3,6 +3,12 @@ name = "ashen"
 version = "0.1.0"
 edition = "2021"
 
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+[lints]
+workspace = true
 
 [dependencies]
+engine = { path = "../engine/" }
+flate2 = "1.0.28"
+
+[dev-dependencies]
+eyre = "0.6.8"
