+++ ashen/src/main.rs
@@ -1,3 +1,5 @@
+mod packfile;
+
 fn main() {
     println!("Hello, world!");
 }
