+++ ashen/src/packfile/mod.rs
@@ -1,10 +1,7 @@
 //! This only works with the last version of Ashen :).
 
-mod nom;
-
+use engine::utils::nom::*;
 use flate2::read::ZlibDecoder;
-#[allow(clippy::wildcard_imports)]
-use nom::*;
 use std::io::Read;
 
 #[derive(Debug, PartialEq)]
