+++ engine/Cargo.toml
@@ -3,6 +3,15 @@ name = "engine"
 version = "0.1.0"
 edition = "2021"
 
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+[lints]
+workspace = true
 
 [dependencies]
+loom = { version = "0.7.1", optional = true }
+nom = "7.1.3"
+
+[dev-dependencies]
+eyre = "0.6.8"
+
+[features]
+loom = ["dep:loom"]
