+++ engine/src/lib.rs
@@ -1,2 +1,3 @@
 pub mod asset;
 pub mod directory;
+pub mod utils;
