--- lib/Cargo.toml
@@ -1,17 +1,0 @@
-[package]
-name = "lib"
-version = "0.1.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-flate2 = "1.0.28"
-loom = { version = "0.7.1", optional = true }
-nom = "7.1.3"
-
-[dev-dependencies]
-eyre = "0.6.8"
-
-[features]
-loom = ["dep:loom"]
