--- lib/src/lib.rs
@@ -1,5 +1,0 @@
-#![allow(unused)] // FIX(Unavailable): just testing :)
-#![warn(unused_imports)]
-#![warn(clippy::pedantic)]
-
-mod packfile;
